Added flow
Desc: We now have flow in the main window, when we
press the home button twice it will scroll up to
the top of the recycler view.
other then that when we press the back button from anywhere isn't home
we return to home screen, if we are already in home we exit the app.